ticket-16 Separate Integration Tests Module
